Reinventing the wheel. Again.
Funny initial git commit messages

Whenever I start a new project, I struggle with what to put in the very
first commit message. After doing a “git init” there is technically
nothing there and no work done yet, so how to describe this initial
step?

“Initial commit” always seemed a little futile, so I started gathering
some better examples on what to put in the message for the glorious
start of a new project and I am sharing them here for all of you to
marvel at:

This is where it all begins...
Commit committed
Version control is awful
COMMIT ALL THE FILES!
The same thing we do every night, Pinky - try to take over the world!
Lock S-foils in attack position
This commit is a lie
I'll explain when you're older!
Here be Dragons
Reinventing the wheel. Again.
This is not the commit message you are looking for
Batman! (this commit has no parents)

Or, you know, just use whatthecommit.com.